- Added ability to declare local attributes by adding current state of to the RyzClass object.
By adding this state object we may keep track if we
are inside a class, a method etc.



- Refactored RyzClass, removed the way we use to
get the class name, now, it depends if we ever matched
a class definition and change the state to "InsideClass"

- Added RyzClassState to keep track of the current state
of the parser.

For instance, if the current state is "initial" it doesn't make sense
to track return values or local assignments.

If we are inside a method ( InsideMethodState ) we
should not care about package or import definitions and so on.

TODO: Re-refactor the State object to avoid high coupling between
the state and the LineTransformers.


added test-resources/04.methods/LocalVariableSpec.ryz