Invoke blocks with "a()", instead of "a.run()".
The interface ryz.lang.Block was added. This class is used to identify block closures.

The change was performed mainly in the class SimpleValueTransformer

A quick and very dirty change was performed to identify if the invoken method
was a block or not. This needs to be changed for a more robust mechanism.