Added extension methods.
Basically anything that fails to compile will be attempted passing receiver as argument to the method.

So, the string reference s when invoked the method reverse()  ( as in s.reverse() ) will be attempt to reverse( s )

Also operations such as  a .+ ( b )  may be solved by these extension methods.

Need more work to handle much other scenarios.